Make it possible to run the RSPM image as a non-root user.
The official Ubuntu packages for RSPM set up an 'rstudio-pm' user to run
the service. However, the existing Docker image could not be run as this
user for two reasons:

* The existing Dockerfile creates some directories required at runtime
  as root; and

* The existing installation step does not use userspace license
  activation [0], so later attempts to activate a license as a non-root
  user will fail.

This commit fixes both these issues. Images still default to running as
root (for now, to preserve compatibility), but they can also be run as
'rstudio-pm' now as well.

[0]: https://docs.rstudio.com/rspm/admin/appendix/minimal-root-use-installation/#licensing

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>