Avoid the COPY + chmod pattern.
Unfortunately, using COPY to move a file into an image followed by RUN
chmod actually duplicates the file and creates an unnecessary image
layer.

This commit makes use of the --chmod flag that was added to COPY for
this exact reason in a few places. This flag is only supported by
BuildKit, but that's what we use to build images anyway, so it seems
like a reasonable win.

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>