fix an issue where the `r-session-complete` image name is computed incorrectly
Most of our images have an `rstudio-` prefix. However, `r-session-complete` does not.
As a result, we create an `$IMAGE_PREFIX` variable to contain the conditional prefix. This
will also prove useful if/when we decide to change the image prefix for most images. We also
removed a lot of double quotes (") and opted for `${}` syntax instead.