add a sleep to workbench startup... 😭
We discussed this offline. Workbench apparently has a race condition where it
starts up and provisions files. If it shuts down before complete, you
can end up with failing tests and a weird `launcher.pem` file with no data
and the wrong file permissions. Hopefully this "fix" will not be long lived.