ci: update readme path for content-base and content-pro dockerhub repos
`content/base/README.md` and `content/pro/README.md` files no longer exist,
so this change updates the update-readme workflow to use the `content/README.md`
file for the dockerhub repository description for the `content-base` and `content-pro`
images.

The intent of this change is to fix the `no such file or directory` error,
which is one of the two errors causing the update-readme workflow to fail.

The next step to get the update-readme workflow to run successfully
is to ensure the dockerhub access token that is used by the
`peter-evans/dockerhub-description` github actions has the `read/write/delete` scope.

Signed-off-by: Lucas Rodriguez <lucas.rodriguez@posit.co>